renames GlobalContextProviders to ApplicationRoot and moves toaster in there
puts back commented out code
